fix: cleanup and move gnodev into /cmd
